Convert Board from struct to class
One of the arguments for falling back to using a class instead
of a struct is that "Copying or comparing instances doesn't make sense",
which can be seen in this StackOverflow anser: https://stackoverflow.com/questions/24232799/why-choose-struct-over-class. I don't think it makes sense
to copy or compare two Game objects, so I will for the time being
convert it to a class until I can come up with a satisfactory answer as
to why it should be a struct.