Support partial coordinates
It was not practical to only support full coordinates: Partial coordinate are practical
when you cannot immediately figure out what source coordinate is meant. Because of
when one disambiguates different pieces, you might only have a file or a rank. So supporting
partial coordinates allows when to start out with nothing and add more information to it
as you discover more.