Move disambiguation of pieces at the start of validation
We did disambiguate pieces, but not before the en passant validation,
which would return the first piece that could perform the en passant,
even though some other piece was actually intended to be used.
Specifically, if two pawns had one column between them and an captured
pawn moved from the opposite side and to between them, then either
of the pawns could capture it, but in this case the move would have to be
disambiguated with the file of the correct pawn.